requirements.txt
alembic==1.5.8
altair==4.2.0
anyio==3.3.4
arcgis==1.8.5.post3
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.6.0
Flask==2.0.3
GeoAlchemy2==0.8.5
geographiclib==1.52
geopy==2.1.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lerc==0.1.0
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
ntlm-auth==1.5.0
numpy==1.21.3
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.20.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyarrow==7.0.0
pycparser==2.20
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.0
pyshp==2.1.3
pytesseract==0.3.9
python-certifi-win32==1.6
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
sacremoses==0.0.49
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
Shapely==1.7.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sodapy==2.1.0
SQLAlchemy==1.3.24
streamlit==1.8.1
terminado==0.12.1
testpath==0.5.0
tokenizers==0.11.6
toml==0.10.2
tomli==1.2.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
transformers==4.18.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
ujson==4.2.0
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wrapt==1.13.2
zipp==3.8.0
